+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a history of the changes made to this project.
 * ADS-B Exchange script now sets up mlat-client to connect to their mlat-server.
 * Added the ability to disply either the dump1090-fa or dump1090-mutability map.
 * Changed the bash function which retrieves config file variables so it works properly.
+* Fixed issue causing a package installation error when PiAware versions change.
 
 ## v2.1.0 *(August 30th, 2016)*